+++ CHANGELOG.md
@@ -17,10 +17,13 @@ Types of changes:
 
 ## [Unreleased]
 
+### Added
+
+* Improved documentation of the postprocessor
+
 ### Fixed
 
 * Name shadowing of `category` (local and `unicodedata` import)
-
 
 ## [0.0.4] - 2024-11-18